requirements_old.txt
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.5.1
Bashutils==0.0.4
billiard==3.5.0.3
certifi==2017.4.17
chardet==3.0.4
cheroot==5.8.3
CherryPy==11.0.0
click==6.7
decorator==4.0.11
Flask==0.12.2
Flask-Cache==0.13.1
Flask-Excel==0.0.6
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
flask-mongoengine==0.9.3
Flask-PyMongo==0.5.1
Flask-QRcode==0.10.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
geetest==3.2.1
gevent==1.2.2
greenlet==0.4.12
httplib2==0.10.3
idna==2.5
image==1.5.5
infinity==1.4
intervals==0.8.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
kombu==4.1.0
lml==0.0.1
lxml==4.1.1
MarkupSafe==1.0
msgpack-python==0.4.8
olefile==0.44
pathlib2==2.3.2
pbr==3.0.1
Pillow==4.1.1
pluggy==0.7.1
portend==2.2
pycrypto==2.6.1
pycurl==7.43.0
pytest==3.7.1
pytest-timeout==1.3.1
python-dateutil==2.6.1
pytz==2017.2
qrcode==5.3
requests==2.18.4
salt==2017.7.2
scandir==1.9.0
singledispatch==3.4.0.3
six==1.10.0
speaklater==1.3
Tempita==0.5.2
tempora==1.9
texttable==0.9.1
urllib3==1.21.1
validators==0.12.0
versioning==0.2.3
vine==1.1.4
virtualenv==15.1.0
Werkzeug==0.12.2
WTForms==2.1
WTForms-Alchemy==0.16.2
WTForms-Components==0.10.3
xlrd==1.0.0
xlwt==1.2.0
